Create gate-tempest-dsvm-neutron-placement-full-ubuntu-xenial-nv
This adds a job to nova's experimental queue for testing changes
with the placement-api service running.

It also runs full tempest with neutron (since nova-network is
deprecated) and on an ubuntu xenial node (which will keep it from
running on stable branches also, which is good).

Depends-On: I53dd3e6b41de17387a0e179fc9ac64c143b6a9eb

Part of blueprint generic-resource-pools

Change-Id: Ifae72c0d686faea859b037ac88c15322959b3b33