require CLA for release-tools
We missed a case in testing releases because the release-test repo did
not require having a CLA signed. The release-tools and release-test
repos share an ACLs file. Rather than split the ACLs, just update the
release-tools ACLs in place to add the flag so that future tests more
accurately reflect what real repositories will look like.

Change-Id: I13b5c5db18888e6d65e9bd8e9ed9feeba66ecffb
Signed-off-by: Doug Hellmann <doug@doughellmann.com>