Run with cells v2 in placement and neutron grenade jobs
cell0 is optional in newton but we support it, and in order
to do that we need some test coverage. We want to test both
ways, with and without cell0 (cellsv2), and by default devstack
and grenade run without it.

This change updates two jobs that run in the nova check queue
that will give coverage to running with cells v2 and the cell0
database.

Nova currently has 3 grenade jobs in it's check queue, but only
one of those runs with neutron, and since cellsv2 is the future
and nova-network is deprecated, we add cells v2 to the single-node
neutron grenade job that runs in nova's check queue.

As for the non-grenade job, we add cellsv2 to the placement API
job which runs with neutron and is non-voting in the nova check
queue. Since the placement-api is optional in newton, and cell0
is optional in newton, it seems natural to lump these into the
same job.

This depends on the devstack and grenade changes, respectively,
that use the new NOVA_CONFIGURE_CELLSV2 flag in devstack.

Depends-On: Ic7d0115da51d6ea17ee49071af259a7789c62ab9
Depends-On: I090d5994c3077d86c60457e4c16a03d35e0c409c

Change-Id: Iacb01bca5d946a260f7d1c397d7f14b00a74bb08