Ensure that gerrit projects have zuul pipelines
Add a layout check to find projects added to gerrit but not zuul, and
vice versa.

We can ignore gerrit projects with a read-only state since they are not
under active development. We can also ignore the z/tempest zuul pipeline
since it is a meta-pipeline not tied to a specific project.

Change-Id: I1da184ecc0268689081c1795b8ee741b41d80d45