do not fail if there are no constraint updates to push
We have been assuming that everything that looks like a library appears
in the upper constraints list and so there would be an update to
review. When that assumption is false the commit step exits with an
error that causes the job to appear to fail, even though it has
not. Intercept the error and report that we are skipping the step
instead of exiting with an error.

Change-Id: Ie185b86735bd91d566a1017c6c98073779ba71f7
Signed-off-by: Doug Hellmann <doug@doughellmann.com>