Add merge commit ACL for packaging-deb
As discuss on IRC, we should have a long therm goal to use the overlay
mode of git-buildpackage. Though it's already very late in the Newton
release and the team still hope to release Newton using OpenStack infra.
Switching to the overlay mode of git-buildpackage would imply:
1/ Rewriting some of the build script
2/ Somehow use git filter-branch on all repositories, in order to clean
everything.

None of the actions above can be achieve in a reasonable timeframe.
Therefore, the only option we have so that the packaging-deb can move
forward, is to allow "merge commits" for the moment, so we can merge
upstream tags matching a new release, until we can switch to the overlay
mode of git-buildpackage.

Change-Id: Ied8519607fcbe08494daaebf05ea980d120bab0d