do not error if a branch already exists
When we run batch operations to create a lot of stable branches at one
time, sometimes we have to run the script more than once and sometimes
we mistakenly run it against the same repo more than once. It's safe to
ignore the existing branch, so don't treat it like an error.

Change-Id: I60dd4d0a30da0ca5d1913c309f24010ab0d446dd
Signed-off-by: Doug Hellmann <doug@doughellmann.com>