Change Kolla jobs to 90 minutes for short term
We are struggling to get mirrors added to openstack-infra because
I believe it is a manual proces and reviews are stalling out waiting
on internal openstack-infra training.  I realize its the end of the
cycle and everyone in openstack-infra is working 18x7, so rather then
fix this properly now (with mirrors which solves the problem) I'd like
to extend our gate timer to 90 minutes for our centos jobs which are
much slower becaue of a lack of mirrors for our dependent software
packages.

Removing bindep install helped 2 minutes - but that isn't enough.  The
reason this has reared its ugly head is likely because we have twice
as many containers in Newton as we had in Mitaka - which puts the build
time at 45 minutes in the gate.

I will follow this up with a stacked review of mirroring for our
repos in Kolla for CentOS (where the gates are failing) in early
Ocata.

Change-Id: If4fbbd84e49e851f3ddac48595a8d8b72ff69db5