fix logic for adding new release notes pages
Start by moving the steps to their own script so they are easier to test
locally. Then fix the logic for inserting new entries into the toctree
so it takes the indentation into account. This version assumes we're
standardizing on listing release notes in reverse chronological order,
and those changes will be made separately.

Change-Id: Ib1f74e05c36ba1d247c324c276cfc2d8d75d9420
Signed-off-by: Doug Hellmann <doug@doughellmann.com>