upstream: and options: track-upstream for deb-*
The usual workflow for packaging a new upstream release, is to do
"git merge -X theirs <TAG>", with <TAG> being the latest upstream
rleease. With Gerrit, this translates into a merge commit.

In order to be able to do merge commits, the upstream code must be
present in the Gerrit repository, otherwise, this translates into
a serries of patch, which is absolutely not what we want.

Therefore, we need to add this to all deb-* projects:

upstream: <URL>
options:
  - track-upstream

This is what this patch does, with specific upstream URL for each
package.

Change-Id: I6c03f90ffc447203a7b842e5b1840463a9123d1f