Create ansible-role-bindep project
We've be creating an empty project to manage bindep. We'll also be
writing an ansible plugin so you can read / install the packages from
bindep.txt.

This project is not part of big tent right now. We'd like irc
notifications to be part of #windmill.

We'll be adding voting jobs for linters / docs, but non-voting
functional tests.

Change-Id: I0e5f68edf9cdeae405ccffafb69e53301b43b13f
Signed-off-by: Paul Belanger <pabelanger@redhat.com>