Include marker file for release notes
This is so that we can create a list of directories to ignore when we
rsync documentation builds to a publishing site. Some builds are
published as subdirectories underneath the results of superior builds.
This allows those superior builds to be rsynced without removing the
other builds that were published under them.

Change-Id: Ie2eb32c4853188fa059018dffb0e161fcf95c085