Make subunit2sql py35 jobs voting
This commit switches the subunit2sql py35 jobs to be voting.
Subunit2sql has supported running on py35 for a long time (since
archlinux updated their default python version to use it) The nv drop
was just delayed by a mysql 5.7 incompat which was introduced by the
switch to xenial.

Change-Id: Id104c1cbd03069bb80f11f407d79210ec57a307e