Temporary use master for requirements in translation job
The stable/newton branch is not cut yet - and translation jobs for
stable/newton fail. Add a temporary mapping from stable/newton to master
so that translation sync work and translators can work on the branch.

Note the two builders are only used by translation jobs and not anywhere
else. So, this should only affect translation jobs.

Example fail:
http://logs.openstack.org/89/8938a26d6b5f747717649f55c761775e3cc52ad3/post/heat-upstream-translation-update/ddf9fae/

Change-Id: Ie30f9fc9976d95804dd8d83ec8a15271b286af54