Move Puppet OpenStack gate to Puppet 4 voting
We decided to start voting on Puppet 4 starting from Newton and keep
Puppet 3 jobs only for stable branches.

This patch removes experimental jobs for puppet4, to move them in the
gate, and configure Zuul to run Puppet 3 and 4 jobs only for their
respective branches.

Note: it also removes the multinode job that we never finished and we
don't need it anymore.

Change-Id: I00cee8d2356993cb9ad006329afcdc05c2fad9f1