puppet: do not run puppet3 xenial jobs on stable mitaka/liberty/hammer
In a previous patch I updated Zuul Layout to reflect our desire to run
Puppet 4 tests in the gate instead of Puppet 3 for current master and
future stable/newton.
In the change, I forgot to update some other rules that were applied
later in the layout.

This patch does 3 things:
- Bring all layout related to puppet jobs together, so it's easier to
  read.
- Do not run Puppet 3 Xenial jobs, we don't want it. We started to
  support Puppet4 from now, we want to gate on it.
- Run Puppet 4 for experimental scenario004 job. Puppet 3 is not
  wanted anymore.

Change-Id: Ib5a4c0f8282412e824410c30e97d85529577b2ad