Check changed track-upstreams in gerrit/projects.yaml
Check changes to the upstream URL for a repo that has option
track-upstream. The check currently is just cloning of the repository.

Since we currently have over 580 repos with track upstream, this check
compares the previous version of the file with the current one and only
tests for changed entries.

Change-Id: Ic716c685512ebe660876388d663314bf3b008b66