Add JJB AFS module
This adds a JJB module that validates the parameters we use for
the afs publisher.  This publisher is not an actual Jenkins
publisher and only exists in Zuul v2.5's ansible launcher.  Thus,
it has no XML output and is defined locally in this repo.

This also removes the run-layout script (which used to clone HEAD
of both jjb and zuul) in favor of installing these dependencies
via pip in the tox testenv.  We are no longer tracking JJB master,
so pinning to the version of JJB we are using makes more sense.

Similary, it is not expected that we will need cutting edge
versions of zuul for this test in the immediate future (we may
want to revisit this when we approach zuulv3).

Change-Id: I4acd70c868cb6965d39f968b7582257b4ab21778