Stop running compare-xml jobs
This removes the compare-xml jobs from both the project-config
and the jenkins-job-builder projects.  The JJB configuration in
project-config is beginning to include custom JJB components
which are supported only in Zuul v2.5.

For the project-config directory, we install a custom JJB module
to support these extensions and run JJB to validate that we have
a good configuration.

It would be possible to do the same for changes to JJB, but
considering that we are beginning to use custom components while
pinning to a released version of JJB (rather than master), it
seems like the project-config corpus of jobs may be less relevant
now, so perhaps now is a good time to stop.

Change-Id: I7c96b866663ce497b566a20f55c4c715b13c3723