Use the DIB image id in glance uploads
In order to make use of shade's intelligence in not reuploading an image
that is already uploaded with the same hash value we need to make sure
that any reattempts at uploads use the same image name. To do this we
move away from timestamp suffixes in the glance image names and instead
use the DIB image id. This way if we try to upload the same DIB image to
a cloud more than once shade should notice and make this very very
cheap.

Change-Id: Ie6c82942ecd7d27e45e9092c137660d016d996c5