Add jobs to networking-cisco for testing version compat
networking-cisco is aiming to be compatible with multiple versions of
neutron. The project has custom tox environments for running the unit
tests against particular versions of newton and global requirements.
This adds jobs to run those custom tox environments on each patch, with
the end goal being that we prevent merging code that doesn't maintain
the multi-version compatibility we are after.

Change-Id: I605885f6216a99d31fbd342b7c4e8427a765e984
Depends-On: I2932ebc263bf13265c63fdd6896feecf9e09581d