[Manila] Add lost job to master and newton branches pipelines
Change [1] added xenial jobs to Manila project and three of existing
voting jobs on trusty should have still run for master and
stable/newton branches until xenial ones become stable.
But in [1] was lost one of filters for one of voting jobs that was
enabling it for all branches. So, add it to have full stack again.

[1] I8af1340f3cb415a7211e13122aea503a2981a1da

Change-Id: I98ff9d0014f9ebc22b4ab3871342933eac829849