Retire some wrong deb-os-* and deb-oslo*
These repositories are duplicate of deb-python-* (which matches the
source package name uploaded to Debian). They were wrongly added to
packaging-deb.

This patch is the intial step to remove the packages as per the
instructions at http://docs.openstack.org/infra/manual/drivers.html.

Our appologies for the time loss with this.

Change-Id: Ie5dab06b847490340669bdd09c494f09a8e81e87