Add TripleO multinode in Nova experimental jobs
TripleO jobs are useful to test deployments outside Devstack and would
provide interesting feedback to Nova developers:
- is a patch backward compatible outstide Devstack?
- is a feature working with a default deployment?

This job is non-voting and is run at demand with experimental pipeline.
It's a first iteration of improving Nova (and OpenStack in general)
coverage by tools used by OpenStack operators.

By having this job in place, TripleO team will see if tooling is working
well to build Nova package in CI jobs and see if the job is actually
working fine and useful.

Change-Id: I51f54061dd13b36e49ed3e07c75c7050722191df