Rework ironic-lib tempest jobs
Now that most of our provisioning logic lives in ironic-lib, we need
the same jobs running on it as on ironic itself. This change creates
a similar template to ironic's and replace one existing job with
generated jobs.

Switch master and newton to xenial, while leaving mitaka on trusty.
Switch jobs to ipmitool as it's planned anyway.

Change-Id: I83f9e5f3e01772ca59b1ec303e83858148e9c2a6