Use --version-map when generating constraints updates
We need to generate constraints for all python versions we use.  This is
python 2.7 (trusty and xenial), 3.4 (trusty) and 3.5(xenial).  The best
way to do that is run generate-constraints with all 3 versions of python
and have it sort out the freeze.

This is difficult as there isn't a trivial way to have all 3 versions of
python available, and building from source on the proposal node is too
intrusive.

Use a --version-map that allows us to clone the output of a freeze
between python versions.  This assumes that the python version in
question are compatible during the transition window.

This will mean that we generate constraints consistently between
versions regardless of which version /usr/bin/python3 is.

Change-Id: I1a0122407a1edd8b828465780b3e7bd0d1b719d3
Depends-On: I55a3597ea2f2bf48e30dc564a4bf4b2ef7f178b3