Improve release note translation import criteria
Release notes are usually translated per release.
For example, translators want to focus on Newton release notes.
Currently translation progress is checked as a whole, so translators
need to translate other releases too so that translation for a
specific release is published.

It looks better to check release notes translation progress
per release. This commit updates the propose translation script
to add a PO file if translation at least for one release is
translated well (>=75%) and remove a PO file when translations
for all releases have poor progress (<40%).

Change-Id: I10d057866a9d063fe51da57b454189c587268c5d