Skip CentOS7 jobs for OSA OpenStack service repositories
Skip any OpenStack-Ansible jobs for os_ repositories which run on
CentOS7 except os_keystone, os_glance, os_cinder for stable/newton.

CentOS7 support is ongoing. While many repositories have CentOS7
support and jobs to validate this, we need to skip these jobs for
most of the OpenStack (ie os_*) repositories as they do not yet have
the supporting code in stable/newton but are gaining it in master.

Change-Id: I7f74834fd87ed07fcec86bf9dec2ea20ea5c5b48