Extend job time for coverage comparisons
The networking-cisco project has a tox "cover" target that is tied to the
coverage job. This tox target does a coverage check of baseline code, and then
coverage check of the change-set. The reports are compared to determine if
coverage is better or worse for a pass/fail decision for the job.

This job has been timing out on some commits. After discussion with infra [1],
a new coverage job (compare-cover) is being proposed, and will use a 75 minute
timeout. We may be able to reduce this in the future, based on run results.

Made the job template with gate-{name}-compare-coverage-{node}, so that it can
be used on other projects. For networking-cisco, which is now branchless, the
job gate-networking-cisco-compare-coverage-ubuntu-xenial will be used. Commits
for networking-cisco will be on master, and will use the ubuntu-xenial node.

Note: networking-cisco tox.ini has been modified, under [2], to associate the
target compare-cover, with the script that performs this comparison based
coverage test.

[1] https://goo.gl/J3Y7YV
[2] https://review.openstack.org/#/c/385575

Change-Id: Ifcc754055d8d1749fd24e707232948ca731980e6