temporarily configure mistral to publish to pypi
We have a few projects that import Mistral code as though it was a
library. Those projects rely on being able to run unit tests through
tox, which wants to pip install Mistral. We have several very old
versions of the project on PyPI, but not the new Newton releases. We
will tag a new Newton release using the same SHA as the most recent
patch release to trigger publishing after this patch is merged. When the
work to create mistral-lib is done, we can revert this patch.

Change-Id: I9fb8be010b2b9c5466c8b293fdc72763323ad2dd
Needed-by: I1c98b9f120da81caeb27750f419b3712aeefaaff
Signed-off-by: Doug Hellmann <doug@doughellmann.com>