Create new trove CI jobs
Part of the project [1] to get rid of the trove-integration repository
(in the Queens release), and to get rid of the tool redstack is to
fold into the trove project, all the elements and image building
components that were part of trove-integration.

To facilitate an easy transition, a multi-step approach was
chosen. Now that the "stuff" from trove-integration is in the trove
branch, it is time to turn on the CI on these things, and begin to
switch to using these as the voting CI for the trove project in Ocata
and forward, leaving the old trove-integration based approach for the
stable branches.

The trove-integration based approach will continue to be used for
Newton, and that will be history when the Queen release comes out. If
we can turn on the CI on trove that doesn't use trove-integration (for
master) we will be on the way to that.

This change accomplishes the following:

- rename gate-trove-buildimage-{datastore} to
  gate-trove-integration-buildimage-{datastore}{job-suffix}
- create new gate-trove-buildimage-{datastore}{job-suffix} jobs
- rename gate-trove-functional-* to gate-trove-legacy-functional-*
- create new gate-trove-functional-* job
- rename gate-trove-scenario-functional-dsvm-* to
  gate-trove-legacy-scenario-functional-dsvm-*
- create new gate-trove-scenario-functional-dsvm-* job
- fix branch override and datastores for gate-trove-functional (now
  gate-trove-legacy-functional)

The newly created jobs are experimental, and non-voting and as soon as
they are shown to be working OK, will be flipped into the check (and
gate if their legacy predecessors were voting) to replace the legacy
jobs for master.

[1] https://review.openstack.org/#/c/373153/

Change-Id: I5ff0f00052db18931946fc3b9618d6b5e52c5720