Fix Newton branch regex for some releasenotes
Change I3d2007f053d2a8fd62b1d183b7747ea66d42d949 updated the branch
regex to include newton - but the repo updated in that change
(python-tripleoclient) is not part of the branch condition.

The change of the branch regex is wrong and not mentioned in commit
message. Revert it. This is also in accordance with the commit message.

Instead add a new branch regex for releasenotes starting with Ocata.
This follows the second changeset of
I3d2007f053d2a8fd62b1d183b7747ea66d42d949.

Change-Id: I87179e2d547421bc85fbf2a2abff233e467f2180