rationalize trove CI jobs for stable branches
As part of the project to get rid of the trove-integration repository,
all things there were moved into the trove repository early in
Ocata. Therefore, for stable liberty/mitaka and newton, CI will
continue to use trove-integration, and for master (soon to be Ocata),
and beyond we don't use trove-integration.

In refactoring the CI jobs, I failed to handle the stable branches
properly and for two jobs in particular, gate-trove-functional-dsvm-
and gate-trove-scenario-functional-dsvm-, I did not ensure that stable
branches got the right jobs. This commit aims to fix that.

I hope :)

Here is the detailed list of changes being made.

1. in jenkins/jobs/projects.yaml, many of the trove-integration jobs
were mixed in with the trove jobs in the trove project. Made a
trove-integration project and moved the trove-integration specific
jobs there.

2. in jenkins/jobs/projects.yaml the trove job
gate-trove-legacy-functional-dsvm- is now independent of the
trove-integration job by the same name.

3. in zuul/layout.yaml added branch specifiers for the
gate-trove-legacy- and gate-trove- non-legacy jobs.

4. in zuul/layout.yaml, add gate-trove-legacy-scenario-functional-
jobs to trove.

Change-Id: I8bfdc22b814799b2aaaa7ecd077bb2e5a9b989e4