Make the cells v1 job run explicitly with nova-network
The cells v1 job won't work with neutron, so it has to run
with nova-network. As we're looking to globally switch
DEVSTACK_GATE_NEUTRON to 1 (the default) in devstack-gate
for the master branch (Ocata) if not specified, we need to
specify this for the cells v1 job.

See I6ece30e823c0cf21fbb2f37568c669dd7d3c6a08

Change-Id: If86b061be095ed5cfc7a5f5594f84abae68baab0