update trove ci for python-troveclient project
This change adds a skip on the test gate-trove-legacy-functional-* for
branches other than liberty, mitaka and newton.

It extends the logic that skips this for the trove project. I tried to
have two projects in the same skip-if but that didn't seem to work;
this one does seem to work.

Change-Id: I7a96bfead5b9cc69ebd27848b269d4f6353ba263