Explicitly set DEVSTACK_GATE_NEUTRON=0 for nova-net grenade jobs
We're going to have conditional logic in devstack-gate for setting
the default value of DEVSTACK_GATE_NEUTRON. That will only set
the value from a default if not explicitly defined, and it's going
to change the default to enable neutron in Ocata.

This change explicitly sets DEVSTACK_GATE_NEUTRON=0 in the
nova-network specific grenade jobs.

Note there are some other grenade jobs defined which don't specify
if neutron or nova-network should be used so they always relied
on the default in devstack-gate. So for grenade going from newton
to ocata those jobs will start using neutron, which is a change
and might work, or might not, but we'll have to explicitly define
the network backend to use, or cap those jobs based on branch, if
they don't work with neutron.

Change-Id: I33e8c81bb5366d2d1feccc144baa54ed3ed51e33