remove wheel jobs from release queue
We do not need to run all of the wheel jobs every time we tag a
release, so remove them from that queue. We leave them in the periodic
queue for now, with the idea that we will move them to a higher
frequency in a separate patch, if needed.

Change-Id: I400c5c3df1229a9dd010babad539452e3e1404ab
Signed-off-by: Doug Hellmann <doug@doughellmann.com>