Use dummy share driver of manila with rally CI jobs
For the moment we use two different share back ends for two existing
Rally CI jobs that run against manila project. It is redundant, as in
case of CI we intend to test main manila functionality, that is
back end agnostic. So, switch them to 'dummy' share driver and
disable all other redundant services, that also will reduce deployment
time.

Change-Id: I753a648349f7a3b8ede3ee714657638f4db68ec5