Make gate-rally-dsvm-neutron-neutron voting in the neutron check queue
The decision stems from the fact that the job has been stable for some
time. Also, more scenarios are being added (e.g. [1]), and if momemtum
keeps up, it would be nice to ensure that the job stays green over
time.

Absence of the job on the gate queue is not an oversight. This is a
conscious decision the team is willing to make for the following
reasons:

a) rally runs can be expensive: every scenario may easily add 5-10
   more minutes to the whole runtime. This can have a bad compounding
   effect on the entire gate pipeline for integrated projects.
b) performance degradation may occasionally show up with the combination
   of two or more effects, the same way an intermittent failure may
   show up only after merge, however in fashion that may be less
   obvious to troubleshoot. To avoid crippling the entire gate until
   the issue is pinpointed and resolved, it is best to keep the effects
   of a potential instability confined to the neutron project only.

[1] https://review.openstack.org/#/c/392265/

Change-Id: I65c570ad0eb8aed6632dd0328b21904cc2cc008a