Require explicit list of publishers for each job
Publisher are required for almost every job and it's very easy to forget
to specify them and get a job that cannot be debugged. If publishers are
not required for a job, one should explicitely specify empty list of
publishers as I did here for test job and for rtfd hook.

Change-Id: I7c4923910b0e139ca3831b6450b0fbb82b618b64