neutron: move xenial rally job into check queue
Since there are no visible issues with the job, and it immediately
passed on the very first experimental check in neutron master, I suggest
we just replace the existing trusty job with it, making the latter
voting without the usual time to collect statistics.

If you wonder why the job is in check queue only, it reflects the
current setup for trusty flavour.

This patch still keeps trusty job flavour for Mitaka and earlier.

Change-Id: Iecb6d063cd99257b1d5f640943c743a207fce51f