Restrict gate-novaclient-dsvm-functional <= Newton
The gate-novaclient-dsvm-functional job requires nova-network
which we won't be running except in cells v1 job configurations
in Ocata. This restricts the job from running on changes that
aren't in newton or older. On master the neutron variant of this
job gets sufficient coverage as the nova-network CLIs/APIs in
nova-network were deprecated in newton.

Change-Id: I12d2108f4a82722e839c636b136fd4ca055f7828