Remove gate-neutron-lbaasv2-dsvm-minimal from neutron check queue
This job feels like an exception as no other neutron subproject has
veto over neutron patches. It was added to provide a weather bell
for potential breaking changes, but quite some time has gone since
then and we have become a lot better at spotting/dealing with bumps
in the road. It is only fair to treat lbaas as any other project and
take this job off the neutron queue.

Change-Id: I0ef4df6d48874003fcd94899a43d8e8bc2a7b63d