Ironic: Switch postgres job to standard PXE
This patch is disabling iPXE in favor of testing the standard PXE with
the postgres job. Ironic has a lot of configuration permutations and our
gate already contains a lot of jobs so it makes sense to merge some jobs
that can be be tested together, postgres + standard PXE in this case.

Change-Id: Iffdd369f9224727ae9ae5a08d104a0faa5efc48a