Move devstack updown test to Xenial
The devstack "updown" test is still running on Trusty, making it the
defacto devstack/trusty tester ... a role it is not well suited for.
When this test works, it cleans up after itself and thus deletes all
it's logs, making it hard to compare anything when something goes
wrong.

A follow-on change will propose a specific Trusty job for this

Change-Id: If05224a2d5d880a2affc8d0f2ed5e0b8f554f022