Get rid of the trove-image-builder project
The project was to store image elements for Trove. These elements are
now in the trove repository and therefore branched with the project,
used in the CI, and so on. In short, we don't need this independent
repository any longer.

Process that was followed is as in [1].

In particular, "End Project Gating"[2]: This project never got far
enough to have gating checks and therefore zuul/layout.yaml always
looked like this. It was never in jenkins/jobs/projects.yaml.

Project content was removed in [3], thanks Andreas.

This commit accomplishes [4]. The project is delisted from
zuul/layout.yaml, the project is marked read-only. The project is
removed from channels.yaml.

This completes all the steps of the retirement process.

In addition, [5] is a change to get rid of the project in the
governance repository.

I found that the documentation [1] does not document that a change is
required in gerrit/projects.yaml. I've pushed up a change[6] for that.

[1] http://docs.openstack.org/infra/manual/drivers.html#retiring-a-project
[2] http://docs.openstack.org/infra/manual/drivers.html#end-project-gating
[3] https://review.openstack.org/#/c/399697/
[4] http://docs.openstack.org/infra/manual/drivers.html#remove-project-from-infrastructure-systems
[5] https://review.openstack.org/#/c/398692/
[6] https://review.openstack.org/401459

Change-Id: I4186340ef13cdbaa2df0a752b68e2e4bd6968494
Depends-On: I7d5e82c038f757a23df3cb0b036b86541fe11930
Related: I7a1aef279d1f90c1be4be11fe042db0db1d12ea0