Skip ubuntu-trusty jobs for OpenStack-Ansible on Ocata+
OpenStack-Ansible is removing support for ubuntu-trusty in master
(Ocata) moving forward. We still need the gate jobs to run against
previous branches (stable/newton, stable/mitaka, liberty).

This patch adjusts the conditional that selects hosts for ubuntu-trusty
jobs, and ensures it only runs against stable/newton and earlier
branches.

Additionally, this patch promotes the xenial AIO job to be voting and
part of the gate for Ocata. This is required for the gate to have a
voting job.

Finally, the xenial non-voting job is skipped by Ocata, in favour of the
voting job. This is to ensure the non-voting job is still run by
stable/newton (which skips the voting job).

Change-Id: I1b39df23ae6e166c6eefb5461b60e5dcfb8f5c8a