Fix OpenStack-Ansible xenial gate jobs
With https://review.openstack.org/#/c/400197/ the Trusty gates were
removed for master, and the Xenial gate promoted to voting for master
(but not for Newton or before). Additionally, the doc-guide gate job had
a typo which is fixed here.

The regex was slightly wrong, so the xenial gate is running twice on
master and stable/newton (once as non-voting, once as voting). This PR
aims to fix that.

Change-Id: I94c4b37723664790e1c9868473c64905452b6679