specify node type in gate-trove-scenario-dsvm- jobs
This change updates the trove CI
gate-trove-scenario-dsvm-{datastore}{job-suffix} jobs and adds a
-{node} between the datastore and the job suffix.

All existing jobs are carried forward as is with a node type of
ubuntu-trusty. New non-voting jobs are defined for each such job with
a node type of xenial.

The changes to get trove's gate working with xenial are now almost
complete in [1] and the job created for that
(gate-trove-functional-{datastore}-{node}{job-suffix}) now passes
reliably. ie. we are able to run the functional test suite on xenial
and that implies that the image building is working on the xenial
node. This change gives us scenario tests for node type of
xenial. Once these jobs all pass, a subsequent change will flip the
xenial jobs to voting and remove the trusty jobs completing the
migration to xenial.

[1] https://review.openstack.org/#/c/393221/

Change-Id: I18c74486b4e130e99eddfa25bb43835634bc1b24