Merge install and deploy guide scripts
The two scripts do the same, just have one different parameter. Use a
single script and rename it.

The scripts are included directly in yaml and do not life anymore in the
scripts directory so that changes to them will be directly and don't
need new images to be built.

Change-Id: Ia229b118daabb1f43da1507a37a4fd3724c2512a