TripleO: Test Barbican in scenario002
As discussed in the mailing list [1], the scenario tests will be
moved to the tripleo-heat-templates which will allow us to have
branch-based tests. So now we can have Barbican again in scenario002
as it was previously done.

We also aim to match the scenarios that are tested in the puppet CI,
so getting Barbican to this scenario matches that as well.

[1] http://lists.openstack.org/pipermail/openstack-dev/2016-November/107816.html

Change-Id: I34aa1e5f8bb8953a0832a813f6485318c815fa58