Remove trusty jobs from Neutron master and Newton branches
As all identified issues with Ubuntu Xenial has been addressed and
ubuntu-xenial jobs ran successfuly in the check queue, let's replace
trusty with xenial on Newton and master branches.

The aim of this patch is to leave functional-ubuntu-trusty and
fullstack-ubuntu-trusy running on  stable/mitaka
branch while stable/newton and master branches should be running on
-ubuntu-xenial jobs. All fullstack jobs should remain non-voting at this
point.

Depends-On: Ic5419afe7170e759749afd7055441c82c317efe0
Change-Id: I8c4ac0c9bfef27ce8b1b57df818764b48fdfb16f