trove-integration only runs for stable branches now
Jobs for trove-integration are only to be run for stable
branches. This is because trove-integration (unbranched) is to be
eliminated soon. To that end, all the stuff in trove-integration was
moved into the branched trove repository early in the ocata cycle (now
called master). However, that change was not done for liberty, mitaka
and newton.

Therefore, when making a change to trove-integration, don't run any
jobs against the master branch of trove (hence the jobs being deleted
in this change).

Only run the jobs against the stable branches. Since liberty is going
bye-bye anyway, that job has been eliminated as well. We've made the
last release of liberty that we are going to be making and therefore,
no reason(s) exist to continue to test against that branch.

Change-Id: I2bcaf5cc6b67c103a88244e1c3ba955ef887547c