oslo.messaging: use non-dsvm for functional job
All functional jobs don't need a full devstack vm
but only some packages to be installed

So we can leverage the bindeps.txt feature to just install
our package requirements.

This change adds new jobs to use a non-dsvm.

Once the new jobs are working we will switch them to voting
and remove the dsvm one.

Change-Id: I9a1a997299f33f9c7dc2b8683b03081781a2d5d6