networking-odl: remove periodic tempest liberty job
It turned out that
periodic-tempest-dsvm-networking-odl-{beryllium, boron}-snapshot-liberty
fail to run test. Since stable/liberty branch is about to be end of
life, there is no point to fix it, but remove those periodic jobs.
note: periodic-tempest-dsvm-networking-odl-boron-snapshot-mitaka
is able to run test and failed with some test cases.
It's another story to address it and to add new periodic jobs
with newton branch + new ODL version.

Change-Id: I5587f97a4b893c84492d01fe7f5d3efd068b41ab